==> 04_data_analysis/008_pseudotime/06_scenic_results.qmd <==
---
title: SCENIC results
execute:
  eval: true
code-fold: true
---

# Load packages

```{r}
source(here::here(
  "04_data_analysis/990_code_libraries/02_library-calls-for-renv.R"
))
Sys.setenv(TAR_PROJECT = here::here())
tar_config_set(store = here::here("_targets"))
library(pheatmap)
library(circlize)
library(tradeSeq)
```

```{r}
source("04_data_analysis/998_paper_figures/00_figures_source.R")
```

# Load data 

```{r}
# Get seurat object
tar_load(sce)
tar_load(gene_ids)
tar_load(translated_id)

assertthat::assert_that(
  sum(rownames(sce@assays$RNA@counts) == gene_ids$ensembl) == nrow(gene_ids)
)
```

```{r}
# read scenic results
scenic <- read.csv(here(
  "03_data/990_processed_data/008_pseudotime/02_scenic_results/auc_mtx.csv"
))
scenic[1:3, 1:4]
dim(scenic)

# add metadata
meta <- sce@meta.data |>
  rownames_to_column("Cell")

# check apoe distribution
meta |>
  dplyr::distinct(donor, .keep_all = TRUE) |>
  dplyr::count(diagnosis, apoe_status)

# Start by updating celltype labels to abbreviated ones
celltype_abbrev_df <- readr::read_tsv(here(
  "05_figures/990_shared_figures",
  "003_final_figures",
  "001_data_for_plots/celltype_abbreviations.tsv"
))

meta <- meta |>
  left_join(
    celltype_abbrev_df,
    by = join_by("subcelltype_annotations" == "celltype")
  ) |>
  dplyr::mutate(
    celltypes_abbrev = if_else(
      is.na(celltypes_abbrev),
      subcelltype_annotations,
      celltypes_abbrev
    )
  )

res <- scenic |>
  dplyr::left_join(meta, by = "Cell")

auc <- scenic |>
  column_to_rownames("Cell") |>
  as.matrix()
auc[1:3, 1:4]
regulons <- colnames(auc)
assertthat::assert_that(all(rownames(auc) == res$Cell))
```

```{r}
meta_min <- meta |>
  dplyr::select(Cell, celltypes_abbrev, diagnosis)
```

```{r}
auc_df <- auc |>
  as.data.frame() |>
  rownames_to_column("Cell") |>
  left_join(meta_min, by = "Cell")

# Pivot into long format
auc_long <- res |>
  pivot_longer(
    cols = all_of(colnames(auc)),
    names_to = "Regulon",
    values_to = "AUC"
  )
```


```{r}
# Average AUC per regulon per cell type in controls
mean_auc <- auc_long |>
  dplyr::filter(diagnosis == "Control") |>
  group_by(celltypes_abbrev, Regulon) |>
  summarise(median_AUC = median(AUC), .groups = "drop")
```


```{r}
# Pivot to wide for comparisons
auc_wide <- mean_auc |>
  pivot_wider(
    names_from = celltypes_abbrev,
    values_from = median_AUC
  )

# Filter for candidate regulons
candidate_regulons <- auc_wide |>
  dplyr::filter(Capillary > 0.03 & Arterial < 0.05 & EndoMT_1 > 0.03)
```


```{r}
regulons_of_interest <- candidate_regulons$Regulon

regulons_of_interest <- c(
  "MAFK...",
  "CREM...",
  "SMAD1...",
  "JDP2...",
  "ZFP1..."
)

plot <- auc_long |>
  # Remove unneded celltypes
  dplyr::filter(!grepl("SMC", celltypes_abbrev)) |>
  dplyr::filter(Regulon %in% regulons_of_interest) |>
  ggplot(aes(x = subcelltype_annotations, y = AUC, fill = diagnosis)) +
  geom_violin(scale = "width", trim = TRUE) +
  facet_wrap(~Regulon, scales = "free_y") +
  theme_classic(base_size = 14) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(x = "Cell type", y = "Regulon AUC")

ggsave(
  plot = plot,
  file = here(
    "05_figures/990_shared_figures/004_pseudotime",
    "02_scenic/capillary_regulons.pdf"
  ),
  width = 7,
  height = 10
)
```

Make heatmap of average AUC per celltype and across diagnosis

```{r}
mean_auc_diag <- auc_long |>
  dplyr::mutate(diagnosis = if_else(diagnosis == "Case", "AD", diagnosis)) |>
  group_by(celltypes_abbrev, diagnosis, Regulon) |>
  summarise(mean_AUC = mean(AUC), .groups = "drop")

auc_heatmap <- mean_auc_diag |>
  pivot_wider(
    names_from = c(celltypes_abbrev, diagnosis),
    values_from = mean_AUC
  )

heatmap_mat <- as.data.frame(auc_heatmap)
rownames(heatmap_mat) <- heatmap_mat$Regulon
heatmap_mat <- heatmap_mat[, -1] # drop Regulon column
heatmap_mat <- as.matrix(heatmap_mat)


pheatmap(
  heatmap_mat,
  scale = "row",
  clustering_method = "ward.D2",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  border_color = NA,
  fontsize_row = 8
)


pdf(
  here(
    "05_figures/990_shared_figures/004_pseudotime/02_scenic/all_regulons.pdf"
  ),
  width = 7,
  height = 26
)
pheatmap(
  heatmap_mat,
  scale = "row",
  clustering_method = "ward.D2",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  border_color = NA,
  fontsize_row = 8
)
dev.off()

auc_heatmap <- mean_auc_diag |>
  dplyr::filter(mean_AUC > 0.05) |>
  pivot_wider(
    names_from = c(celltypes_abbrev, diagnosis),
    values_from = mean_AUC
  )

heatmap_mat <- as.data.frame(auc_heatmap)
rownames(heatmap_mat) <- heatmap_mat$Regulon
heatmap_mat <- heatmap_mat[, -1] # drop Regulon column
heatmap_mat <- as.matrix(heatmap_mat)

heatmap_mat[is.na(heatmap_mat)] <- 0


pheatmap(
  heatmap_mat,
  # scale = "row",
  clustering_method = "ward.D2",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  border_color = NA,
  fontsize_row = 8
)

pdf(
  here(
    "05_figures/990_shared_figures/004_pseudotime/02_scenic/auc_filter_0.05.pdf"
  ),
  width = 7,
  height = 18
)
pheatmap(
  heatmap_mat,
  clustering_method = "ward.D2",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  border_color = NA,
  fontsize_row = 8
)
dev.off()
```

Make a heatmap of just the regulons of interest

```{r}
auc_heatmap <- mean_auc_diag |>
  dplyr::filter(
    Regulon %in%
      c("MECOM...", "FOXP2...", "BCL6...", "LEF1...", "ARID5B...", "SOX13...")
  ) |>
  dplyr::mutate(
    # remove anything after "..."
    Regulon_clean = str_remove(Regulon, "\\.\\.\\..*"),
    RegulonDiag = paste(Regulon_clean, diagnosis, sep = "_")
  )
mean_auc_diag |>
  dplyr::filter(grepl("^M", Regulon))

# Step 2: Reshape
df_wide <- auc_heatmap |>
  dplyr::select(celltypes_abbrev, RegulonDiag, mean_AUC) |>
  pivot_wider(names_from = RegulonDiag, values_from = mean_AUC)

# Step 3: Make matrix
mat <- df_wide |>
  column_to_rownames("celltypes_abbrev") |>
  as.matrix()

qs::qsave(
  mat,
  here(
    "05_figures/990_shared_figures/003_final_figures/001_data_for_plots/scenic_tf_matrix.qs"
  )
)

# Extract clean regulon names and diagnosis
col_names_split <- str_match(colnames(mat), "^(.*)_(AD|Control)$")
regulon_labels <- col_names_split[, 2]
diagnosis_vec <- col_names_split[, 3]


# Extract diagnosis from column names
col_diagnosis <- ifelse(grepl("_AD", colnames(mat)), "AD", "Control")

# Update column names in the matrix
colnames(mat) <- regulon_labels

library(circlize)
# Create the heatmap and split columns by diagnosis
manual_column_order <- c(
  "Arterial",
  "Capillary",
  "EndoMT_1",
  "EndoMT_2",
  "Pericyte-2",
  "T-Pericyte",
  "M-Pericyte",
  "Pericyte-1",
  "Arteriolar-SMC",
  "Vascular-SMC-1",
  "Vascular-SMC-2"
)
manual_column_order[manual_column_order %in% rownames(mat)]
ordered_matrix <- mat[manual_column_order, ]
rownames(ordered_matrix) <- str_replace_all(rownames(ordered_matrix), "_", "-")

hmap <- Heatmap(
  ordered_matrix,
  name = "Mean AUC",
  col = colorRamp2(c(0, max(mat)), c("blue", "red")),
  column_split = col_diagnosis, # This is what creates the facets
  column_title_gp = gpar(fontsize = 10),
  cluster_columns = FALSE,
  cluster_rows = FALSE,
  row_names_gp = gpar(fontsize = 9), # Set row label text size
  column_names_gp = gpar(fontsize = 8),
  height = unit(13, "cm"),
  heatmap_legend_param = list(
    direction = "horizontal",
    labels_gp = gpar(fontsize = 7),
    title_gp = gpar(
      fontsize = 8,
      fontface = "bold"
    ),
    # title_position = "leftcenter",
    gap = unit(0.1, "mm"),
    grid_height = unit(1.5, "mm")
  ),
  show_column_names = TRUE
)
# height = unit(3, "mm"),

draw(hmap)
pdf(
  here(
    "05_figures/990_shared_figures/004_pseudotime/02_scenic/regulons_of_interest_heatmap.pdf"
  ),
  width = 5,
  height = 7
)
draw(hmap)
dev.off()
```

# tradeseq plotsmoothers

Want to make a plot of the relevant TF expression to go along with the heamtap

```{r}

sce_subset <- qs::qread(here(
  "03_data/990_processed_data/008_pseudotime",
  "slingshot_tradeseq_3k_case_and_control.qs"
))

counts <- assays(sce_subset)$counts

# Get stats for the all genes overall across all lineages and conditions
plot_gene_pseudotime <- function(gene, gene_symbol, sce, counts) {
  plot <- plotSmoothers(
    sce,
    counts,
    gene,
    curvesCols = pal_npg("nrc")(4),
    size = 0.2,
    lwd = 0.8
  ) +
    scale_colour_npg(
      labels = c(
        "EndoMT-SMC - AD",
        "EndoMT-SMC - Control",
        "EndoMT-PC - AD",
        "EndoMT-PC - Control"
      )
    ) +
    ggtitle(if_else(gene_symbol == "", gene, gene_symbol)) +
    labs(colour = "Lineage") +
    theme(legend.position = "none")
  return(plot)
}

genes <- translated_id |>
  dplyr::filter(
    hgnc_symbol %in%
      c(
        "FOXP2",
        "BCL6",
        "MECOM",
        "LEF1",
        "ARID5B",
        "SOX13"
      )
  )

plot_genes <- map2(
  unique(genes$ensembl_gene_id),
  unique(genes$hgnc_symbol),
  plot_gene_pseudotime,
  sce_subset,
  counts
)

plot_genes_cleaned <- map(
  plot_genes,
  ~ .x +
    theme(title = element_text(size = 8))
)
plot <- plot_grid(plotlist = plot_genes_cleaned, ncol = 2)
```

```{r}
legend_b <- get_legend(
  plot_genes_cleaned[[1]] +
    guides(color = guide_legend(nrow = 2, override.aes = list(size = 5))) +
    labs(colour = "") +
    theme(legend.position = "bottom")
)
p2 <- grid.grabExpr(
  draw(hmap, heatmap_legend_side = "bottom"),
  ht_gap = unit(100, "mm")
)
right_plot <- plot_grid(plot, legend_b, ncol = 1, rel_heights = c(1, 0.1))
key_genes_plot <- plot_grid(p2, right_plot, ncol = 2)
key_genes_plot

ggsave(
  file = here::here(
    "05_figures/990_shared_figures/003_final_figures",
    "scenic_tfs.pdf"
  ),
  plot = key_genes_plot,
  width = 7,
  height = 7,
  units = "in"
)
```
